refactor: amalgamate FormSections & EditorsContent files
